+++ Controls-demo/Filter/Button/ChooseDate.wml
@@ -1,3 +1,3 @@
-<Controls.dateRange:LiteSelector
+<Controls.dateRange:RangeShortSelector
    emptyCaption="Январь'18"
 />
