+++ Controls-demo/Input/Date/RangeLinkLite.wml
@@ -3,15 +3,15 @@
       <div class="PageBlock__title"></div>
       <div class="PageBlock__content">
          <div class="PageBlock__componentWrap">
-            <Controls.dateRange:LiteSelector
+            <Controls.dateRange:RangeShortSelector
                attr:test_name="default_range">
                <ws:monthTemplate>
                   <ws:partial periods="{{_periods}}" template="Controls/shortDatePicker:MonthTemplate"/>
                </ws:monthTemplate>
-            </Controls.dateRange:LiteSelector>
+            </Controls.dateRange:RangeShortSelector>
          </div>
          <div class="PageBlock__componentWrap ControlsDemo-RangeLinkLite_capture">
-            <Controls.dateRange:LiteSelector
+            <Controls.dateRange:RangeShortSelector
                prevArrowVisibility="{{true}}"
                nextArrowVisibility="{{true}}"
                bind:startValue="_startValueHalfYear"
@@ -20,7 +20,7 @@
             />
          </div>
          <div class="PageBlock__componentWrap">
-            <Controls.dateRange:LiteSelector
+            <Controls.dateRange:RangeShortSelector
                bind:startValue="_startValueBind"
                bind:endValue="_endValueBind"
                chooseHalfyears="{{false}}"
@@ -31,10 +31,10 @@
                <ws:monthTemplate>
                   <ws:partial periods="{{_periods}}" template="Controls/shortDatePicker:MonthTemplate"/>
                </ws:monthTemplate>
-            </Controls.dateRange:LiteSelector>
+            </Controls.dateRange:RangeShortSelector>
          </div>
          <div class="PageBlock__componentWrap">
-            <Controls.dateRange:LiteSelector
+            <Controls.dateRange:RangeShortSelector
                bind:startValue="_startValueQuarter"
                bind:endValue="_endValueQuarter"
                chooseHalfyears="{{false}}"
@@ -46,14 +46,14 @@
 
          </div>
          <div class="PageBlock__componentWrap">
-            <Controls.dateRange:LiteSelector
+            <Controls.dateRange:RangeShortSelector
                viewMode="label"
                emptyCaption="не задан"
 			   attr:test_name="range_not_value"
             />
          </div>
          <div class="PageBlock__componentWrap">
-            <Controls.dateRange:LiteSelector
+            <Controls.dateRange:RangeShortSelector
                bind:startValue="_startValue2"
                bind:endValue="_endValue2"
                chooseHalfyears="{{false}}"
@@ -64,7 +64,7 @@
          </div>
       </div>
       <div class="PageBlock__componentWrap">
-         <Controls.dateRange:LiteSelector
+         <Controls.dateRange:RangeShortSelector
                  bind:startValue="_startValueYear"
                  bind:endValue="_endValueYear"
                  prevArrowVisibility="{{true}}"
