+++ Controls-demo/dateRange/LiteSelector/ValueNotSpecified/ValueNotSpecified.wml
@@ -1,3 +1,3 @@
 <div class="controls-demo_LiteSelector-block">
-    <Controls.dateRange:LiteSelector/>
+    <Controls.dateRange:RangeShortSelector/>
 </div>
