+++ Controls/_dateRange/RangeSelector/_RangeSelector.less
