+++ Controls/_dateRange/RangeShortSelector/_RangeShortSelector.less
