+++ Controls/dateRange.less
@@ -1,6 +1,6 @@
 @import "Controls/_dateRange/Input/_Input";
 @import "Controls/_dateRange/RelationButton/_RelationButton";
-@import "Controls/_dateRange/LiteSelector/_LiteSelector";
-@import "Controls/_dateRange/Selector/_Selector";
+@import "Controls/_dateRange/RangeShortSelector/_RangeShortSelector";
+@import "Controls/_dateRange/RangeSelector/_RangeSelector";
 @import "Controls/_dateRange/LinkView/_LinkView";
-@import "Controls/_dateRange/Link/_Link";
+@import "Controls/_dateRange/DateSelector/_DateSelector";
