+++ Controls/dateRange.ts
@@ -1,11 +1,14 @@
 export {default as Input} from './_dateRange/Input';
 export {default as Selector} from './_dateRange/Selector';
+export {default as RangeSelector} from './_dateRange/RangeSelector';
 export {default as LiteSelector} from './_dateRange/LiteSelector';
+export {default as RangeShortSelector} from './_dateRange/RangeShortSelector';
 export {default as RelationController} from './_dateRange/RelationController';
 export {default as RelationWrapper} from './_dateRange/RelationWrapper';
 export {default as RelationButton} from './_dateRange/RelationButton';
 export {default as Utils} from './_dateRange/Utils';
 export {default as Link} from './_dateRange/Link';
+export {default as DateSelector} from './_dateRange/DateSelector';
 export {default as LinkView} from './_dateRange/LinkView';
 export {default as RangeSelectionController} from './_dateRange/Controllers/RangeSelectionController';
 export {default as DateRangeSelectionController} from './_dateRange/Controllers/DateRangeSelectionController';
@@ -21,8 +24,8 @@ export {IDateRangeSelectable};
  * Библиотека контролов, которые служат для отображения диапазона дат и выбора дат из выпадающей панели.
  * @library Controls/dateRange
  * @includes Input Controls/_dateRange/Input
- * @includes Selector Controls/_dateRange/Selector
- * @includes LiteSelector Controls/_dateRange/LiteSelector
+ * @includes RangeSelector Controls/_dateRange/RangeSelector
+ * @includes RangeShortSelector Controls/_dateRange/RangeShortSelector
  * @includes RelationController Controls/_dateRange/RelationController
  * @includes RelationWrapper Controls/_dateRange/RelationWrapper
  * @includes RelationButton Controls/_dateRange/RelationButton
@@ -32,7 +35,7 @@ export {IDateRangeSelectable};
  * @includes IDateRangeSelectable Controls/_dateRange/interfaces/IDateRangeSelectable
  * @includes IRangeSelectable Controls/_dateRange/interfaces/IRangeSelectable
  * @includes Utils Controls/_dateRange/Utils
- * @includes Link Controls/_dateRange/Link
+ * @includes DateSelector Controls/_dateRange/DateSelector
  * @includes RangeSelectionController Controls/_dateRange/Controllers/RangeSelectionController
  * @includes DateRangeSelectionController Controls/_dateRange/Controllers/DateRangeSelectionController
  *
@@ -44,8 +47,8 @@ export {IDateRangeSelectable};
  * Date range library
  * @library Controls/dateRange
  * @includes Input Controls/_dateRange/Input
- * @includes Selector Controls/_dateRange/Selector
- * @includes LiteSelector Controls/_dateRange/LiteSelector
+ * @includes RangeSelector Controls/_dateRange/RangeSelector
+ * @includes RangeShortSelector Controls/_dateRange/RangeShortSelector
  * @includes RelationController Controls/_dateRange/RelationController
  * @includes RelationWrapper Controls/_dateRange/RelationWrapper
  * @includes RelationButton Controls/_dateRange/RelationButton
@@ -54,7 +57,7 @@ export {IDateRangeSelectable};
  * @includes IPeriodLiteDialog Controls/_dateRange/interfaces/IPeriodLiteDialog
  * @includes IDateRangeSelectable Controls/_dateRange/interfaces/IDateRangeSelectable
  * @includes Utils Controls/_dateRange/Utils
- * @includes Link Controls/_dateRange/Link
+ * @includes DateSelector Controls/_dateRange/DateSelector
  * @includes RangeSelectionController Controls/_dateRange/Controllers/RangeSelectionController
  * @includes DateRangeSelectionController Controls/_dateRange/Controllers/DateRangeSelectionController
  * @includes IRangeSelectable Controls/_dateRange/interfaces/IRangeSelectable
