+++ Controls/interface/IFilterButton.js
@@ -132,13 +132,13 @@ define('Controls/interface/IFilterButton', [
     *       templateName="wml!MyModule/panelTemplate"
     *       items="{{_items}}">
     *       <ws:lineSpaceTemplate>
-    *          <Controls.dateRange:LiteSelector
+    *          <Controls.dateRange:RangeShortSelector
     *             startValue="{{_startValue}}"
     *             endValue="{{_endValue}}"/>
     *       </ws:lineSpaceTemplate>
     *    </Controls.filter:Selector>
     * </pre>
-    * @see Controls.dateRange:LiteSelector
+    * @see Controls.dateRange:RangeShortSelector
     */
 
    /*
@@ -153,13 +153,13 @@ define('Controls/interface/IFilterButton', [
     *       templateName="wml!MyModule/panelTemplate"
     *       items="{{_items}}">
     *       <ws:lineSpaceTemplate>
-    *          <Controls.dateRange:LiteSelector
+    *          <Controls.dateRange:RangeShortSelector
     *             startValue="{{_startValue}}"
     *             endValue="{{_endValue}}"/>
     *       </ws:lineSpaceTemplate>
     *    </Controls.filter:Selector>
     * </pre>
-    * @see Controls.dateRange:LiteSelector
+    * @see Controls.dateRange:RangeShortSelector
     */
 
    /**
