+++ tests/ControlsUnit/Input/Date/Link.test.js
@@ -52,26 +52,6 @@ define([
             sandbox.restore();
          });
       });
-
-      describe('_onResultWS3', function() {
-         it('should generate valueChangedEvent and close opener', function() {
-            const
-               sandbox = sinon.sandbox.create(),
-               component = calendarTestUtils.createComponent(dateRange.Link, options),
-               value = new Date(2018, 11, 10);
-
-            component._children.opener = {
-               close: sinon.fake()
-            };
-            sandbox.stub(component, '_notify');
-
-            component._onResultWS3(null, value);
-
-            sinon.assert.calledWith(component._notify, 'valueChanged');
-            sinon.assert.called(component._children.opener.close);
-            sandbox.restore();
-         });
-      });
       describe('_rangeChangedHandler', function() {
          it('should generate valueChangedEvent', function() {
             const
